install dependencies and change pedometer file
