+++ src/screens/CoinBankScreen/CoinBankScreen.js
@@ -1,12 +1,81 @@
-import React from "react";
-import {
-  Text
-} from "react-native";
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+import { Pedometer } from 'expo-sensors';
+import { firebase } from '../../firebase/config';
 
-export default function CoinBankScreen(props) {
-  return (
-    
-      <Text>Hello! This is the book coins page!</Text>
-    
-  )
+export default class PedometerScreen extends React.Component {
+  state = {
+    isPedometerAvailable: 'checking',
+    pastStepCount: 0,
+    currentStepCount: 0,
+  };
+
+  componentDidMount() {
+    this._subscribe()
+  }
+
+  componentWillUnmount() {
+    this._unsubscribe()
+  }
+
+  _subscribe = () => {
+    this._subscription = Pedometer.watchStepCount(result => {
+      this.setState({
+        currentStepCount: result.steps,
+      });
+    });
+
+    Pedometer.isAvailableAsync().then(
+      result => {
+        this.setState({
+          isPedometerAvailable: String(result),
+        });
+      },
+      error => {
+        this.setState({
+          isPedometerAvailable: 'Could not get isPedometerAvailable: ' + error,
+        });
+      }
+    );
+
+    const end = new Date();
+    const start = new Date();
+    start.setDate(end.getDate() - 1);
+    Pedometer.getStepCountAsync(start, end).then(
+      result => {
+        this.setState({ pastStepCount: result.steps });
+      },
+      error => {
+        this.setState({
+          pastStepCount: 'Could not get stepCount: ' + error,
+        });
+      }
+    );
+  };
+
+  _unsubscribe = () => {
+    this._subscription && this._subscription.remove();
+    this._subscription = null;
+  };
+
+  render() {
+      console.log(this.state)
+    return (
+      <View style={styles.container}>
+        {/* <Text>Pedometer.isAvailableAsync(): {this.state.isPedometerAvailable}</Text>
+        <Text>Steps taken in the last 24 hours: {this.state.pastStepCount}</Text> */}
+        <Text>{this.state.currentStepCount}</Text>
+      </View>
+    );
+  }
 }
+
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+  });
+
